+++ library/stub.lua
@@ -488,9 +488,6 @@ local _LFO = {}
 ---@field y2 integer
 local _LineSegment = {}
 
----@class _Menu : playdate.menu
-local _Menu = {}
-
 ---@class _MenuItem : playdate.menu.item
 ---@field title string
 ---@field value (integer|boolean|string)
@@ -991,7 +988,7 @@ function playdate.menu:getMenuItems() end
 --- Removes the specified playdate.menu.item from the menu.
 ---
 --- [Inside Playdate: playdate.menu:removeMenuItem](https://sdk.play.date/Inside%20Playdate.html#m-menu.removeMenuItem)
----@param menuItem _Menu
+---@param menuItem _MenuItem
 ---@return nil
 function playdate.menu:removeMenuItem(menuItem) end